appointment service, and some fixes

+++ migrations/000007_create_service_table.up.sql
@@ -4,13 +4,9 @@ CREATE TABLE IF NOT EXISTS services (
     name varchar(50) NOT NULL,
     description TEXT,
     duration INTERVAL NOT NULL,
-    price DECIMAL(10, 2) NOT NULL,
+    price INTEGER NOT NULL,
     photo_url TEXT,
     created_at TIMESTAMP NOT NULL DEFAULT CURRENT_TIMESTAMP,
     updated_at TIMESTAMP NOT NULL DEFAULT CURRENT_TIMESTAMP
 );
 
-
--- INSERT INTO services (institution_id, name, description, duration, price, photo_url) VALUES
---                                                                                          (1, 'Corte de cabelo', 'Corte de cabelo masculino', '00:30:00', 20.00, 'https://www.google.com.br'),
---                                                                                          (1, 'Corte de cabelo', 'Corte de cabelo feminino', '00:30:00', 30.00, 'https://www.google.com.br');
